requirements/requirements.txt
onnxruntime-genai
numpy
beautifulsoup4
readability-lxml
playwright
httpx
aiofiles

